Move LLM costs from definition to run detail view
- Remove aggregated LLM costs from definition detail (not useful)
- Add per-run LLM cost stats to agent run detail (useful)
- Show token counts and cost in run detail LLM tab
- Simplify API endpoints: /run/{run_id} -> /{run_id}
