+++ props/frontend/src/components/RunDetail.svelte
@@ -501,6 +501,30 @@
         </div>
       {:else if activeLogTab === "llm"}
         <div class="p-4">
+          {#if run.llm_costs}
+            <div class="grid grid-cols-2 md:grid-cols-5 gap-4 text-sm mb-4 pb-4 border-b">
+              <div>
+                <span class="text-gray-500">Requests:</span>
+                <span class="ml-1 font-medium">{run.llm_costs.total_requests.toLocaleString()}</span>
+              </div>
+              <div>
+                <span class="text-gray-500">Input:</span>
+                <span class="ml-1 font-medium">{run.llm_costs.total_input_tokens.toLocaleString()}</span>
+              </div>
+              <div>
+                <span class="text-gray-500">Cached:</span>
+                <span class="ml-1 font-medium">{run.llm_costs.total_cached_tokens.toLocaleString()}</span>
+              </div>
+              <div>
+                <span class="text-gray-500">Output:</span>
+                <span class="ml-1 font-medium">{run.llm_costs.total_output_tokens.toLocaleString()}</span>
+              </div>
+              <div>
+                <span class="text-gray-500">Cost:</span>
+                <span class="ml-1 font-medium text-green-600">${run.llm_costs.total_cost_usd.toFixed(4)}</span>
+              </div>
+            </div>
+          {/if}
           {#if loadingLLMRequests}
             <p class="text-gray-500">Loading LLM requests...</p>
           {:else}
